# openapi-complete.yaml
openapi: 3.1.0
info:
  title: Tazzi Communications Platform API
  version: 1.0.0
  description: |
    Complete API specification for the Tazzi Communications Platform - A unified communications platform
    supporting voice, SMS, email, WhatsApp, and multi-channel messaging.

    ## Authentication

    The API supports two authentication methods:
    - **Bearer JWT Token**: Use `Authorization: Bearer <token>` header
    - **API Key**: Use `X-API-Key: <api-key>` header

    ## Base URL

    Production: `https://api.tazzi.app`

    ## Rate Limiting

    API requests are rate-limited per tenant. Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Time when limit resets (Unix timestamp)

  contact:
    name: Tazzi API Support
    email: support@tazzi.app
    url: https://tazzi.app
  license:
    name: Proprietary
servers:
  - url: https://api.tazzi.app/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /video-calls/rooms:
    post:
      tags: [Video Calling]
      summary: Create a video room
      description: Creates a new video room for instant or scheduled meetings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Room name
                roomType:
                  type: string
                  enum: [instant, scheduled, recurring, webinar]
                  default: instant
                maxParticipants:
                  type: integer
                  default: 10
                maxVideoParticipants:
                  type: integer
                  default: 6
                settings:
                  type: object
                  properties:
                    video_enabled:
                      type: boolean
                    audio_enabled:
                      type: boolean
                    screenshare_enabled:
                      type: boolean
                    recording_enabled:
                      type: boolean
                    waiting_room_enabled:
                      type: boolean
                scheduledAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Room created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  room:
                    $ref: '#/components/schemas/VideoRoom'
    get:
      tags: [Video Calling]
      summary: List video rooms
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [waiting, active, ended]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of video rooms
  /video-calls/rooms/{roomId}:
    get:
      tags: [Video Calling]
      summary: Get room details
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Room details with participants
    patch:
      tags: [Video Calling]
      summary: Update room settings
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                settings:
                  type: object
      responses:
        '200':
          description: Room updated
  /video-calls/rooms/{roomId}/end:
    post:
      tags: [Video Calling]
      summary: End a video room
      description: Ends the room and disconnects all participants
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Room ended successfully
  /video-calls/rooms/code/{roomCode}:
    get:
      tags: [Video Calling]
      summary: Get room by code
      description: Lookup a room by its join code
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room found
        '404':
          description: Room not found or ended
  /video-calls/rooms/{roomId}/invitations:
    post:
      tags: [Video Calling]
      summary: Create invitation
      description: Send an invitation to join the video room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                phone:
                  type: string
                invitedUserId:
                  type: string
                  format: uuid
                message:
                  type: string
                expiresInHours:
                  type: integer
                  default: 24
      responses:
        '200':
          description: Invitation created with join URL
  /video-calls/rooms/{roomId}/recording/start:
    post:
      tags: [Video Calling]
      summary: Start recording
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                recordingType:
                  type: string
                  enum: [composite, individual]
                format:
                  type: string
                  default: mp4
                resolution:
                  type: string
                  default: 1280x720
      responses:
        '200':
          description: Recording started
  /video-calls/rooms/{roomId}/recording/stop:
    post:
      tags: [Video Calling]
      summary: Stop recording
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recording stopped
  /video-calls/rooms/{roomId}/recordings:
    get:
      tags: [Video Calling]
      summary: Get room recordings
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of recordings
  /video-calls/recordings/{recordingId}/download:
    get:
      tags: [Video Calling]
      summary: Get recording download URL
      parameters:
        - name: recordingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Presigned download URL
  /video-calls/settings:
    get:
      tags: [Video Calling]
      summary: Get video settings
      responses:
        '200':
          description: Tenant video settings
    patch:
      tags: [Video Calling]
      summary: Update video settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                videoEnabled:
                  type: boolean
                maxConcurrentRooms:
                  type: integer
                maxRoomParticipants:
                  type: integer
                defaultVideoQuality:
                  type: string
                  enum: [low, sd, hd, fhd]
                recordingEnabled:
                  type: boolean
      responses:
        '200':
          description: Settings updated
  /video-calls/analytics:
    get:
      tags: [Video Calling]
      summary: Get video usage analytics
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Video usage analytics
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from /api-keys
...
  schemas:
...
    VideoRoom:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        roomCode:
          type: string
          description: Unique 9-character join code (XXX-XXX-XXX)
        roomType:
          type: string
          enum: [instant, scheduled, recurring, webinar]
        status:
          type: string
          enum: [waiting, active, ended, cancelled]
        maxParticipants:
          type: integer
        maxVideoParticipants:
          type: integer
        settings:
          type: object
          properties:
            video_enabled:
              type: boolean
            audio_enabled:
              type: boolean
            screenshare_enabled:
              type: boolean
            recording_enabled:
              type: boolean
            waiting_room_enabled:
              type: boolean
            require_authentication:
              type: boolean
        scheduledAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
        createdAt:
          type: string
          format: date-time
...